pyproject.toml
[project]
name = "cachorro-louco-notify"
version = "0.1.0"
description = ""
authors = [
    {name = "mrrobnote",email = "daviesaraujo@hotmail.com"}
]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "discord (>=2.3.2,<3.0.0)",
    "python-dotenv (>=1.2.1,<2.0.0)",
    "requests (>=2.32.5,<3.0.0)",
    "sqlalchemy (>=2.0.45,<3.0.0)",
    "pandas (>=2.3.3,<3.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
